Module constance::kernel::cfg

Static configuration mechanism for the kernel

Structs

CfgBuilder
A kernel configuration being constructed.

CfgEventGroupBuilder
Configuration builder type for EventGroup.

CfgHunkBuilder
Configuration builder type for Hunk.

CfgInterruptHandlerBuilder
Configuration builder type for InterruptHandler.

CfgInterruptLineBuilder
Configuration builder type for InterruptLine.

CfgStartupHookBuilder
Configuration builder type for StartupHook.

CfgTaskBuilder
Configuration builder type for Task.

CfgTimerBuilder
Configuration builder type for Timer.

DefaultInitTag
As a generic parameter of CfgHunkBuilder, indicates that the hunk should be initialized with Init.

InterruptHandlerTable
A table of combined second-level interrupt handlers.

ZeroInitTag
As a generic parameter of CfgHunkBuilder, indicates that the hunk should be zero-initialized.

Traits

HunkIniter
Implemented on DefaultInitTag and ZeroInitTag when T can be initialized in this way.

Type Definitions

InterruptHandlerFn
A combined second-level interrupt handler.