Struct constance::kernel::Hunk

pub struct Hunk<System, T: ?Sized> { /* fields omitted */ }

Represents a single hunk in a system.

Hunks are nothing more than static variables defined in a kernel configuration. They come in handy when you are designing a component that can be instantiated by a kernel configuration and wanting each instance to have its own separate state data.

Relation to Other Specifications: None. The need for programmatically defining static regions has been traditionally fulfilled by code generation and preprocessor-based composition. The closest thing might be cell internal variables in the component system that comes with the TOPPERS kernels.

Implementations

impl<System: Port, T: ?Sized> Hunk<System, T>

pub fn build() -> CfgHunkBuilder<System, T, DefaultInitTag>

Construct a CfgTaskBuilder to define a hunk in a configuration function.

impl<System, T: ?Sized> Hunk<System, T>

pub unsafe fn transmute<U: ?Sized>(self) -> Hunk<System, U>

Reinterpret the hunk as another type.

Safety

Similarly to core::mem::transmute, this is incredibly unsafe.

The byte offset and length must be valid for the destination type.

impl<System: Kernel, T: ?Sized> Hunk<System, T>

pub fn as_bytes_ptr(this: Self) -> *const [u8]

Get a raw pointer to the raw bytes of the hunk.

pub unsafe fn as_bytes(this: Self) -> &'static [u8]

Get a reference to the raw bytes of the hunk.

Safety

The result might include uninitialized bytes and/or interior mutability, so it might be unsafe to access.

impl<System: Kernel, T: 'static> Hunk<System, T>

pub fn as_ptr(this: Self) -> *const T

Get a raw pointer to the hunk's contents.

impl<System: Kernel, T: 'static> Hunk<System, [T]>

pub fn as_ptr(this: Self) -> *const [T]

Get a raw pointer to the hunk's contents.