Module constance::time

Temporal quantification for the Constance RTOS.

Duck Guy: “Maybe time's just a construct of human perception, an illusion created by—”

Tony the Clock: “(alarm sounds) Eh! Eh! Eh! Eh! EH! EH! EH! EH! EH! EH! EH! EH!”

— Don't Hug Me I'm Scared 2 - TIME

Structs

Duration
Represents a signed time span used by the API surface of the Constance RTOS.

Time
Represents a timestamp used by the API surface of the Constance RTOS.

TryFromDurationError
Error type returned when a checked duration type conversion fails.