Module constance::utils::intrusive_list

Intrusive doubly linked list backed by a container implementing std::ops::Index.

Structs

Ident
Reference wrapper that implements PartialEq and Eq by identity comparison.

Iter
An iterator over the elements of ListAccessorCell.

Link
Links to neighbor items.

ListAccessorCell
Cell-based accessor to a linked list.

ListHead
Circualr linked list header.

Static
A virtual container of Ts that can be indexed by Ident<&'static T>.

Traits

CellLike

Type Definitions

StaticLink
Links to neighbor items with a 'static lifetime.

StaticListHead
Circualr linked list header where elements are linked by StaticLink<Element> (a pair of &'static Element).